change models attributes.

+++ src/main/java/com/d4d/controller/DoctorController.java
@@ -7,6 +7,7 @@ package com.d4d.controller;
 
 import com.d4d.model.Doctor;
 import com.d4d.service.DoctorService;
+import com.d4d.controller.util.ApiKey;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -26,21 +27,25 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/doctor")
 public class DoctorController {
-    //for Token based authentication
-    private String apiKey = "1234";
-
+    
+//for Token based authentication
+    @Autowired
+    private ApiKey apiKey;
+    
     @Autowired
     private DoctorService doctorService;
 
     @ResponseBody
-    @RequestMapping(value = {"/create"}, method = RequestMethod.POST, consumes = "application/json")    
+    @RequestMapping(value = {"/create"}, method = RequestMethod.POST, consumes = "application/json")
     public ResponseEntity<String> create(@RequestHeader("apikey") String api_key, @RequestBody Doctor doctor) {
 
-        if (api_key.equals(apiKey)) {
+        if (apiKey.checkApiKey(api_key)) {
 
             try {
+
                 Doctor p = doctorService.create(doctor);
                 return new ResponseEntity<>("Doctor inserted successfully with id: " + p.getId(), HttpStatus.OK);
+
             } catch (Exception e) {
                 e.printStackTrace();
                 return new ResponseEntity<>("error when adding new Doctor", HttpStatus.EXPECTATION_FAILED);
@@ -52,31 +57,91 @@ public class DoctorController {
     }
 
     @RequestMapping("/get")
-    public Doctor getDoctor(@RequestParam String name) {
-        return doctorService.getByName(name);
+    public ResponseEntity<Doctor> getDoctor(@RequestHeader("apikey") String api_key, @RequestParam String name) {
+        if (apiKey.checkApiKey(api_key)) {
+            try {
+
+                Doctor doctor = doctorService.getByName(name);
+                return new ResponseEntity<>(doctor, HttpStatus.OK);
+
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @RequestMapping("/getAll")
-    public List<Doctor> getAll() {
-        return doctorService.getAll();
+    public ResponseEntity<List<Doctor>> getAll(@RequestHeader("apikey") String api_key) {
+
+        if (apiKey.checkApiKey(api_key)) {
+            try {
+                List<Doctor> doctors = doctorService.getAll();
+                return new ResponseEntity<>(doctors, HttpStatus.OK);
+
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @ResponseBody
     @RequestMapping(value = {"/update"}, method = RequestMethod.POST, consumes = "application/json")
-    public String update(@RequestHeader("apikey") String api_key, @RequestBody Doctor doctor) {
-        Doctor p = doctorService.update(doctor);
-        return p.toString();
+    public ResponseEntity<String> update(@RequestHeader("apikey") String api_key, @RequestBody Doctor doctor) {
+
+        if (apiKey.checkApiKey(api_key)) {
+            try {
+
+                Doctor p = doctorService.update(doctor);
+                return new ResponseEntity<>(p.toString(), HttpStatus.OK);
+
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @RequestMapping("/delete")
-    public String delete(@RequestParam String name) {
-        doctorService.delete(name);
-        return "Deleted " + name;
+    public ResponseEntity<String> delete(@RequestHeader("apikey") String api_key, @RequestParam String name) {
+
+        if (apiKey.checkApiKey(api_key)) {
+            try {
+
+                doctorService.delete(name);
+                return new ResponseEntity<>("Deleted " + name, HttpStatus.OK);
+
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @RequestMapping("/deleteAll")
-    public String deleteAll() {
-        doctorService.deleteAll();
-        return "All Deleted";
+    public ResponseEntity<String> deleteAll(@RequestHeader("apikey") String api_key) {
+
+        if (apiKey.checkApiKey(api_key)) {
+            try {
+
+                doctorService.deleteAll();
+                return new ResponseEntity<>("All Doctors Deleted", HttpStatus.OK);
+
+            } catch (Exception e) {
+                e.printStackTrace();
+                return new ResponseEntity<>(HttpStatus.EXPECTATION_FAILED);
+            }
+        } else {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 }
